backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.0
httpx>=0.26.0
youtube-transcript-api>=0.6.1
google-api-python-client>=2.108.0
openai>=1.12.0
PyPDF2>=3.0.1
python-multipart>=0.0.6

